+++ CHANGES.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2020-11-25
+
+### sdm-api-app:0.3.3
+* Fix MissingMethodException in discovery by calling `toString()` on device type
+
 ## 2020-11-02
 
 ### sdm-api-thermostat:0.1.5